requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
pymongo
pydantic
PyMuPDF
pillow
pytesseract
python-multipart
PyPDF2
python-docx
